rename models to model
