+++ tests/test_plugins.py
@@ -7,7 +7,7 @@ from pytest_mock import MockerFixture
 from hydromt.components import __hydromt_eps__ as component_eps
 from hydromt.components.grid import GridComponent
 from hydromt.drivers import __hydromt_eps__ as driver_eps
-from hydromt.models.model import Model
+from hydromt.model.model import Model
 from hydromt.predefined_catalog import __hydromt_eps__ as catalog_eps
